Add media-rss parsing support
The media-rss data is read into new fields (under media-rss-data) in the entry
struct.

Also enhance the parse-feed's :extra option to potentially parse multiple values
for an element name and to apply specified transform functions. This is a
breaking change for users of the :extra option.

New tests reveal that the <author> field doesn't seem to be read properly. This
seems to be a pre-existing problem.